+++ src/integrationTest/java/uk/gov/hmcts/reform/bulkscanprocessor/tasks/BlobProcessorTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.bulkscanprocessor.scheduler;
+package uk.gov.hmcts.reform.bulkscanprocessor.tasks;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.common.io.Resources;
